+++ markout-docusaurus-plugin/settings.gradle.kts
@@ -1,0 +1,3 @@
+pluginManagement {
+    includeBuild("../build-logic")
+}
