+++ settings.gradle.kts
@@ -3,6 +3,7 @@ pluginManagement {
 }
 
 includeBuild("markout-plugin")
+includeBuild("markout-docusaurus-plugin")
 
 include("markout")
 include("markout-markdown")
